remove dead code + fix contactsfield import
