+++ mootiro_maps/apps/main/utils.py
@@ -12,8 +12,6 @@ from random import choice
 from dateutil.parser import parse as dateutil_parse
 from copy import deepcopy
 
-from jsonfield import JSONField
-
 from django import forms
 from django.http import HttpResponse
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
@@ -568,34 +566,3 @@ def get_model_from_table_ref(table_ref):
     return model
 
 
-class ContactsField(JSONField):
-    json_field_defaults = {
-        'address': None,
-        'compl': None,
-        'city': None,
-        'postal_code': None,
-        'phone': None,
-        'facebook': None,
-        'email': None,
-        'twitter': None,
-        'site': None,
-        '---': None,
-    }
-
-    def __init__(self, *args, **kwargs):
-        kwargs.update(default=self.json_field_defaults)
-        super(ContactsField, self).__init__(*args, **kwargs)
-
-    def key_name(self, key):
-        return {
-            'address': _('Address'),
-            'compl': _('Complement'),
-            'city': _('City'),
-            'postal_code': _('Postal Code'),
-            'phone': _('Phone'),
-            'facebook': _('Facebook'),
-            'email': _('E-mail'),
-            'twitter': _('Twitter'),
-            'site': _('Web Site'),
-            '---': '---',
-        }[key]
